# cmd/main/docs/swagger.yaml
basePath: /api
definitions:
  auth_usecase.AuthLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  http_util.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  user_models.UserWithoutPassword:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  user_usecase.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:3000
info:
  contact:
    email: rafael.cavalcante@tallos.com.br
    name: Rafael Cavalcante
  description: Golang server
  termsOfService: http://swagger.io/terms/
  title: KitKit back golang
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica um usuário pela senha
      parameters:
      - description: Authenticate
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/auth_usecase.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Autorizado
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.HTTPError'
      summary: Autentica um usuário
      tags:
      - Auth
  /users:
    get:
      consumes:
      - application/json
      description: Retorna todos os usuários
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user_models.UserWithoutPassword'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.HTTPError'
      summary: Retorna todos os usuários
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Recurso de criação
      parameters:
      - description: Add User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_usecase.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.HTTPError'
      summary: Cria um usuário
      tags:
      - Users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Retorna um usuário pelo ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_models.UserWithoutPassword'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.HTTPError'
      summary: Retorna um usuário
      tags:
      - Users
swagger: "2.0"